% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{rc_format_variables}
\alias{rc_format_variables}
\title{Clean variable types of raw data export via CSV}
\usage{
rc_format_variables(data, dictionary, yesno = "logical")
}
\arguments{
\item{data}{Data exported from REDCap via csv.}

\item{dictionary}{Data dictionary exported from REDCap via csv.}

\item{yesno}{Determine how to return REDCap 'Yes - No' fields; options include 'factor' (default), 'numeric', or 'logical'.}
}
\value{
The REDCap dataset with variable types cleaned.
}
\description{
Clean variable types of raw data export via CSV
}
