% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{rc_read_api}
\alias{rc_read_api}
\title{Title}
\usage{
rc_read_api(url, token)
}
\arguments{
\item{url}{URL for a REDCap database API. Check your institution's REDCap documentation for this address.}

\item{token}{REDCap API token}
}
\value{
A named list containing four dataframes: dd = metadata, evnt = Events, inst = Instrument mappings, rcrd = Records.
}
\description{
Title
}
