% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{rc_read_csv}
\alias{rc_read_csv}
\title{Imported REDCap csv files and combine into single list.}
\usage{
rc_read_csv(folder)
}
\arguments{
\item{folder}{The folder containing the files downloaded from the REDCap
database. The four files required include the Data Dictionary, Events
(under the 'Define my events' tab), Instrument Mappings (under the
'Designate Instruments for My Events' tab), and the raw Record data (under
the 'My Reports & Exports' tab). All files should be saved as .csv.}

\item{ids}{Names of identifiers, for inclusion on all output datasets.}
}
\value{
A named list containing four data frames: dd = metadata, evnt = Events, inst = Instrument mappings, rcrd = Records.
}
\description{
Clean raw .csv data exported from a REDCap database, and export a list of data.frames
}
