% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{rc_tidy}
\alias{rc_tidy}
\title{Tidy list of datasets from REDCap.}
\usage{
rc_tidy(object, ids = NULL, label = FALSE, repeated = "exclude")
}
\arguments{
\item{object}{An named list containing the following data.frames; metadata (dd),
events (evnt), instruments (inst) and records (rcrd).}

\item{ids}{Names of identifiers, for inclusion on all output datasets.}

\item{label}{Add labels to variables. Supply name of labelling package,
\code{Hmisc} or \code{sjlabelled}.}

\item{repeated}{How shall repeated forms be treated in datasets assembled by
event. Options are: exclude, include, or nest. Nest uses `tidyr` package to
collapse by row id.}
}
\value{
A list of dataframes, with variables grouped by event or by data collection form.
}
\description{
Tidy list of datasets from REDCap.
}
