% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{tidy_redcap}
\alias{tidy_redcap}
\title{Tidy datasets imported from REDCap.}
\usage{
tidy_redcap(folder, ids = c("id", "redcap_event_name"))
}
\arguments{
\item{folder}{The folder containing the files downloaded from the REDCap
database. The four files required include the Data Dictionary, Events
(under the 'Define my events' tab), Instrument Mappings (under the
'Designate Instruments for My Events' tab), and the raw Record data (under
the 'My Reports & Exports' tab). All files should be saved as .csv.}

\item{ids}{Names of identifiers, for inclusion on all output datasets.}
}
\value{

}
\description{
Clean raw .csv data exported from a REDCap database, and export a list of data.frames
}
