% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap.R
\name{rc_tidy}
\alias{rc_tidy}
\title{Tidy list of datasets from REDCap.}
\usage{
rc_tidy(object, ids = NULL, label = FALSE, repeated = "exclude")
}
\arguments{
\item{ids}{Names of identifiers, for inclusion on all output datasets.}

\item{label}{Add labels to variables. Supply name of labelling package,
\code{Hmisc} or \code{sjlabelled}.}

\item{repeated}{How shall repeated forms be treated in datasets assembled by
event. Options are: exclude, include, or nest. Nest uses `tidyr` package to
collapse by row id.}

\item{folder}{The folder containing the files downloaded from the REDCap
database. The four files required include the Data Dictionary, Events
(under the 'Define my events' tab), Instrument Mappings (under the
'Designate Instruments for My Events' tab), and the raw Record data (under
the 'My Reports & Exports' tab). All files should be saved as .csv.}
}
\value{

}
\description{
Clean raw .csv data exported from a REDCap database, and export a list of
data.frames
}
